wasi: generic file table (#38)
I would like to suggest a refactor to extract the file tables and most
of the generic logic out of `unix.System` and into a
`wasi.FileTable[File]` type.

The decoupling of the file table operations, and actual invocation of
operations on a file makes it possible to experiment with new
implementations of `wasi.System` which are not wrappers but bottom
layers (like `unix.System` but potentially having a different
implementation). With this model, we can implement only the specialized
parts. It accelerates our ability to create new systems and ensures that
we minimize the risks of seeing divergence of behaviors.

Please take a look and let me know if you would like to see anything
changed!

---------

Signed-off-by: Achille Roussel <achille.roussel@gmail.com>