unix: consistent unix socket addresses (#61)
This PR ensures unix socket addresses are consistent across platforms.
Only Linux, empty unix socket addresses are seen as `@` because the Go
standard library and /x/sys package consider them to be abstract sockets
(when the first byte in the address is a null byte). On other platforms,
the unix socket address is instead an empty string.